+++ backend/src/authentication/constants/cache.ts
@@ -1,0 +1,3 @@
+// cache key = `auth-webdav-${sha256(login + password)}` => UserModel | null
+export const CACHE_AUTH_WEBDAV_PREFIX = 'auth-webdav' as const
+export const CACHE_AUTH_WEBDAV_TTL = 900 as const